font resize problem (closes #13)
On the Nexus 7 virtual device, the menu choices to make the font smaller
instead made the font larger. While debugging and stepping through the code I
saw that the text size returned on line 109 of NoteEdit.java increased over
successive selections of the "Smaller" menu option - the opposite of what it
should have done, but that explained the visual change that I was seeing on
screen.

I found that this behavior was fixed with a change to the setTextSize() call
on line 111 to explicitly set the unit of change.